+++ src/config/index.js
@@ -1,0 +1,6 @@
+export { default as appInformation } from 'config/appInformation';
+export { default as englishContent } from 'config/languages/englishContent';
+export { default as navigationConfig } from 'config/navigationConfig';
+export { default as socialLinks } from 'config/socialLinks';
+export { default as spanishContent } from 'config/languages/spanishContent';
+export { defaultRedirects, routes } from 'config/routesConfig';