+++ src/layouts/index.js
@@ -1,0 +1,2 @@
+export { default as LayoutDefault } from 'layouts/LayoutDefault';
+export { default as NoLayout } from 'layouts/NoLayout';