+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export { default as AppRoute } from 'utils/AppRoute';
+export { default as MainRouteRedirect } from 'utils/MainRouteRedirect';
+export { default as RoleChecker } from 'utils/RoleChecker';
+export { default as RoutesSwitch } from 'utils/RoutesSwitch';