+++ changelog.md
@@ -18,6 +18,7 @@
 -   Feature: Update theme.
 -   Feature: Update button to accept more custom colors.
 -   Feature: Export all components from a single file.
+-   Feature: Export everything from a single file within config, layouts and utils folders.
 
 ### v0.6.2